requirements.txt
# Run this FIRST in PowerShell to clean everything:
# pip uninstall langchain langchain-core langchain-community langchain-text-splitters langchain-huggingface langchain-chroma langchain-google-genai langchain-groq langchain-classic langsmith google-generativeai chromadb sentence-transformers -y

# Then install:
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
langchain-google-genai==2.0.7
langsmith==0.2.11
chromadb==0.5.3
sentence-transformers==3.0.1
huggingface-hub==0.24.6
pdfplumber==0.11.4
pypdf==4.3.1
docx2txt==0.8
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
tabulate==0.9.0
rank-bm25==0.2.2
google-generativeai==0.8.3